Changed alignment values to match old code, and changed sensor range

+++ src/main/cpp/infrastructure/SwerveModule.cpp
@@ -38,7 +38,7 @@ SwerveModule::SwerveModule(
     const int driveMotorCanID,
     const int turningMotorCanID,
     const int turningEncoderPort,
-    const int alignmentOffset) noexcept : m_name{name}
+    const double alignmentOffset) noexcept : m_name{name}
 {
     // Set up onboard printf-style logging.
     std::string logName{"/SwerveModule/"};
